Add explanation bubble when you click on the range indicator on note millstone screen.

+++ ios/Milestones/StandardMilestone.m
@@ -23,9 +23,9 @@
     NSInteger rangeHigh = self.rangeHigh.integerValue;
 
     if (rangeLow < 30 && rangeHigh < 30) {
-        return [NSString stringWithFormat:@"%d to %d days", rangeLow, rangeHigh];
+        return [NSString stringWithFormat:@"%d to %d days", (int)rangeLow, (int)rangeHigh];
     } else if (rangeLow < 365 && rangeHigh < 365) {
-        return [NSString stringWithFormat:@"%d to %d months", rangeLow / 30, rangeHigh / 30];
+        return [NSString stringWithFormat:@"%d to %d months", (int)rangeLow / 30, (int)rangeHigh / 30];
     } else {
         return [NSString stringWithFormat:@"%@ to %@", [self humanReadableDays:rangeLow], [self humanReadableDays:rangeHigh]];
     }
@@ -33,12 +33,12 @@
 
 - (NSString *)humanReadableDays:(NSInteger)days {
     if (days < 30) {
-        return [NSString stringWithFormat:@"%d day%@", days, days == 1 ? @"" : @"s"];
+        return [NSString stringWithFormat:@"%d day%@", (int)days, days == 1 ? @"" : @"s"];
     } else if (days < 365) {
         int months = (int) (days / 30.5F);
         return [NSString stringWithFormat:@"%d month%@", months, months == 1 ? @"" : @"s"];
     } else {
-        int years = days / 365;
+        int years = (int)days / 365;
         int remainingDays = days % 365;
         int months = (int) (remainingDays / 30.5);
         if (months >= 1) {
